fix(ha-ws-client-go): handle variable HA WebSocket API response types
  - Change HistoryState.LU/LC from int64 to float64 to handle decimal
    Unix timestamps (e.g., 1767306034.233545)
  - Change SysLogEntry.Source from []string to any with GetSource()
    helper to handle variable array element types
  - Fix SubscribeToTemplate race condition by registering event handler
    before sending WebSocket message
  - Fix getHistory() to preserve float64 timestamp precision

  These fixes align Go types with actual Home Assistant WebSocket API
  responses which use variable typing for backwards compatibility.

+++ scripts/ha-ws-client-go/internal/handlers/monitor.go
@@ -302,7 +302,7 @@ func getHistory(ctx *Context, entityID string, timeRange types.TimeRange) ([]typ
 
 		hs := types.HistoryState{}
 		if lu, ok := stateMap["lu"].(float64); ok {
-			hs.LU = int64(lu)
+			hs.LU = lu
 		}
 		if s, ok := stateMap["s"].(string); ok {
 			hs.S = s
